+++ src/main/java/com/picimako/justkitting/intention/callmatcher/CallMatcherGenerator.java
@@ -7,7 +7,7 @@ import java.util.function.Consumer;
 /**
  * Base type for generating {@link com.siyeh.ig.callMatcher.CallMatcher} initializer calls.
  */
-public interface CallMatcherGenerator<T> {
+public interface CallMatcherGenerator<METHOD, METHOD_CALL> {
 
     /**
      * Generates the CallMatcher initializer for the argument method's signature.
@@ -17,5 +17,15 @@ public interface CallMatcherGenerator<T> {
      *                    This makes it possible to execute logic after for example a user's
      *                    choice of some sort on the UI.
      */
-    void generateCallMatcher(T method, Consumer<String> postActions);
+    void generateCallMatcherForMethod(METHOD method, Consumer<String> postActions);
+
+    /**
+     * Generates the CallMatcher initializer for the argument method call's signature.
+     *
+     * @param methodCall the method call to generate a CallMatcher for
+     * @param postActions any action to execute after the CallMatcher initializer is deleted.
+     *                    This makes it possible to execute logic after for example a user's
+     *                    choice of some sort on the UI.
+     */
+    void generateCallMatcherForMethodCall(METHOD_CALL methodCall, Consumer<String> postActions);
 }