+++ src/main/java/com/picimako/justkitting/intention/callmatcher/GenerateCallMatcherFromSignatureIntention.java
@@ -13,16 +13,17 @@ import com.intellij.openapi.project.Project;
 import com.intellij.psi.PsiFile;
 import com.intellij.psi.PsiIdentifier;
 import com.intellij.psi.PsiMethod;
+import com.intellij.psi.PsiMethodCallExpression;
+import com.intellij.psi.PsiReferenceExpression;
 import com.intellij.util.IncorrectOperationException;
 import com.picimako.justkitting.resources.JustKittingBundle;
 import org.jetbrains.annotations.NotNull;
-import org.jetbrains.annotations.Nullable;
 
 import java.awt.datatransfer.DataFlavor;
 
 /**
  * This intention generates a {@link com.siyeh.ig.callMatcher.CallMatcher} initializer call
- * based on the Java class method it is invoked on.
+ * based on the Java class method, or the Java method call, it is invoked on.
  * <p>
  * The generated code is copied to the clipboard, so after pasting it, {@code CallMatcher} has
  * to be imported manually.
@@ -48,25 +49,37 @@ public class GenerateCallMatcherFromSignatureIntention implements IntentionActio
     public boolean isAvailable(@NotNull Project project, Editor editor, PsiFile file) {
         if (!file.getFileType().equals(JavaFileType.INSTANCE)) return false;
 
-        return getMethodSignatureAtCaretOrEmpty(file, editor) != null;
+        return isPsiMethodOrMethodCall(file, editor);
     }
 
-    @Nullable
-    public static PsiMethod getMethodSignatureAtCaretOrEmpty(PsiFile file, Editor editor) {
+    public static boolean isPsiMethodOrMethodCall(PsiFile file, Editor editor) {
         var elementAtCaret = file.findElementAt(editor.getCaretModel().getOffset());
-        return elementAtCaret instanceof PsiIdentifier && elementAtCaret.getParent() instanceof PsiMethod
-            ? (PsiMethod) elementAtCaret.getParent()
-            : null;
+
+        if (elementAtCaret instanceof PsiIdentifier) {
+            var parent = elementAtCaret.getParent();
+            return parent instanceof PsiMethod
+                || (parent instanceof PsiReferenceExpression && parent.getParent() instanceof PsiMethodCallExpression);
+        }
+
+        return false;
     }
 
     //---- Invocation ----
 
     @Override
     public void invoke(@NotNull Project project, Editor editor, PsiFile file) throws IncorrectOperationException {
-        var targetMethod = (PsiMethod) file.findElementAt(editor.getCaretModel().getOffset()).getParent();
-        new JavaCallMatcherGenerator(project, editor)
-            .generateCallMatcher(targetMethod,
-                callMatcher -> CopyPasteManager.getInstance().setContents(new SimpleTransferable(callMatcher, DataFlavor.stringFlavor)));
+        var elementAtCaret = file.findElementAt(editor.getCaretModel().getOffset());
+        if (elementAtCaret.getParent() instanceof PsiMethod) {
+            var targetMethod = (PsiMethod) elementAtCaret.getParent();
+            new JavaCallMatcherGenerator(project, editor)
+                .generateCallMatcherForMethod(targetMethod,
+                    callMatcher -> CopyPasteManager.getInstance().setContents(new SimpleTransferable(callMatcher, DataFlavor.stringFlavor)));
+        } else {
+            var targetMethodCall = (PsiMethodCallExpression) elementAtCaret.getParent().getParent();
+            new JavaCallMatcherGenerator(project, editor)
+                .generateCallMatcherForMethodCall(targetMethodCall,
+                    callMatcher -> CopyPasteManager.getInstance().setContents(new SimpleTransferable(callMatcher, DataFlavor.stringFlavor)));
+        }
     }
 
     @Override