+++ CHANGELOG.md
@@ -4,6 +4,13 @@
 
 ## [Unreleased]
 
+## [0.2.0]
+### Changed
+- Added missing `@Override` annotation to `getState()` and `loadState()` methods when converting a class to `PersistentStateComponent`.
+- Relaxed the conditions in what the service `getInstance()` generation action is available.
+It is now available in interfaces, abstract classes and non-light-service classes as well, and users can also choose the service level.
+- [#8](https://github.com/picimako/just-kitting/issues/8): `CallMatcher` initializers can now be generated when invoked on method calls as well.
+
 ## [0.1.0]
 ### Added
 - Added project files.